feat(adapters): add general 64-bit reduction
Let the scalar backend reduce residues modulo any prime q <= 2^63 - 1 without
a native 128-bit integer type.

- Add a two-limb Barrett path that forms the full 128-bit product and
  reduces it with mu = floor(2^128 / q) kept as two words, computed via
  a UINT64_MAX division plus binary long division.
- Extend Montgomery REDC with an R = 2^64 radix for the general modulus
  range alongside the existing R = 2^32 fast path.
- Dispatch scalar arithmetic on q limb width (single-word Barrett below
  2^32, two-limb above) and stop silently upgrading q >= 2^32 to
  Montgomery.
- Fix the never-assigned inverse-of-N used to normalize the inverse NTT.
- Widen the scalar_toy reference adapter to full 64-bit residues with a
  portable shift-and-add mulmod.

- Factorize q-1 for the complete uint64_t domain with a deterministic
  Miller-Rabin test and Pollard's rho, so primitive NTT roots can be
  derived for any 64-bit prime, not only the known-moduli table.

+++ src/adapters/scalar_toy/ntt_scalar_toy_internal.h
@@ -26,35 +26,35 @@
  * @brief Private state owned by the scalar toy NTT adapter.
  *
  * The common NTT context intentionally does not expose these fields. This
- * adapter uses canonical uint32_t coefficients in [0, q) and computes all
+ * adapter uses canonical uint64_t coefficients in [0, q) and computes all
  * modular arithmetic with generic scalar operations.
  */
 typedef struct {
-    uint32_t q;            /* prime modulus */
+    uint64_t q;            /* prime modulus */
     uint32_t n;            /* transform size */
     uint32_t stages;       /* log2(n) number of NTT stages */
-    uint32_t omega;        /* primitive n-th root of unity mod q */
-    uint32_t omega_inv;    /* inverse of omega mod q */
-    uint32_t n_inv;        /* inverse of n mod q */
-    uint32_t psi;          /* primitive 2n-th root of unity mod q */
-    uint32_t psi_inv;      /* inverse primitive 2n-th root of unity mod q */
-    uint32_t *psi_pow;     /* psi_pow[i] = psi^i mod q, i = 0..n-1 */
-    uint32_t *psi_inv_pow; /* psi_inv_pow[i] = psi^-i mod q, i = 0..n-1 */
+    uint64_t omega;        /* primitive n-th root of unity mod q */
+    uint64_t omega_inv;    /* inverse of omega mod q */
+    uint64_t n_inv;        /* inverse of n mod q */
+    uint64_t psi;          /* primitive 2n-th root of unity mod q */
+    uint64_t psi_inv;      /* inverse primitive 2n-th root of unity mod q */
+    uint64_t *psi_pow;     /* psi_pow[i] = psi^i mod q, i = 0..n-1 */
+    uint64_t *psi_inv_pow; /* psi_inv_pow[i] = psi^-i mod q, i = 0..n-1 */
 } ntt_scalar_toy_state;
 
-uint32_t ntt__reduce(uint32_t a, uint32_t q);
-uint32_t ntt__addmod(uint32_t a, uint32_t b, uint32_t q);
-uint32_t ntt__submod(uint32_t a, uint32_t b, uint32_t q);
-uint32_t ntt__mulmod(uint32_t a, uint32_t b, uint32_t q);
-uint32_t ntt__modpow(uint32_t base, uint32_t exp, uint32_t q);
-uint32_t ntt__modinv(uint32_t a, uint32_t q);
+uint64_t ntt__reduce(uint64_t a, uint64_t q);
+uint64_t ntt__addmod(uint64_t a, uint64_t b, uint64_t q);
+uint64_t ntt__submod(uint64_t a, uint64_t b, uint64_t q);
+uint64_t ntt__mulmod(uint64_t a, uint64_t b, uint64_t q);
+uint64_t ntt__modpow(uint64_t base, uint64_t exp, uint64_t q);
+uint64_t ntt__modinv(uint64_t a, uint64_t q);
 
 bool ntt__validate_modulus(const ntt_config *config);
 void *ntt__adapter_setup(const ntt_config *config);
 void ntt__adapter_teardown(void *state);
 
-int ntt__forward(void *state, uint32_t *a);
-int ntt__inverse(void *state, uint32_t *a);
-int ntt__negacyclic_mul(void *state, uint32_t *a, uint32_t *b, uint32_t *c);
+int ntt__forward(void *state, uint64_t *a);
+int ntt__inverse(void *state, uint64_t *a);
+int ntt__negacyclic_mul(void *state, uint64_t *a, uint64_t *b, uint64_t *c);
 
 #endif /* NTT_SCALAR_TOY_INTERNAL_H */
